day15: perf: use vector to store lower numbers
The hashmap lookup time increased significantly while the numbers
increases due to hashmap collision and increased link table lookup.
We also noticed the number of calls are heavily congested on lower
numbers, so this gives us a way to optimize the lookup time without
increase too much space usage.

As the total duty cycle of numbers are about 1/8, we store the lowest
1/10 of numbers in vector, so they are guranteed O(1) lookup time,
for rest of numbers we use hashmap to avoid the unused numbers to
occupy space(in vector case)

This reduced the executing time by 3 times without obvious memory
usage increasing